+++ integrations/nvidia/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [integrations/nvidia-v0.1.1] - 2024-11-14
+
+### 🐛 Bug Fixes
+
+- Fixes to NvidiaRanker (#1191)
+
 ## [integrations/nvidia-v0.1.0] - 2024-11-13
 
 ### 🚀 Features